+++ doc/changelog.rst
@@ -4,11 +4,16 @@
 Changelog
 =========
 
-This document describes the main changes in TRIQS 1.0 that might
-have consequences for your scripts and archives.
+This document describes the main changes in TRIQS.
+
+From TRIQS 0.x to TRIQS 1.0
+---------------------------
+
+There have been changes from versions 0.x to 1.0 that will most likely have
+consequences for your scripts and archives.
 
 Python classes
---------------
+~~~~~~~~~~~~~~
 
 The number of python classes in the old 0.x releases was increasing with no
 clear naming convention. In TRIQS 1.0 we have unified the naming of the classes
@@ -17,9 +22,17 @@ following the `PEP naming conventions
 
 * Package and module names: lowercase with underscores
 * Class names: CapWords
+* Function names: lowercase with underscores
 * Function and method arguments: lowercase with underscores
 
 Archives
---------
+~~~~~~~~
 
+We provide :download:`an update script <scripts/update_archive.py>` which should
+help you upgrade your archive. The usage is very simple::
 
+  $ python update_archive.py old_archive new_archive
+
+where ``old_archive`` is your old archive to be upgraded and ``new_archive`` is
+the name of the new archive. If you encounter problems please post an
+issue with a copy of your archive.